requirements.txt
pydantic
fastapi
uvicorn
python-dotenv
numpy
Pillow
pyzbar
requests
tensorflow
onnxruntime
torch
transformers
urllib3==1.26.14